Migration fails when using SQL Server
The issue is with the migration SQL scripts. When it comes to SQL Server
you can execute the ALTER and UPDATE on the same table in the same
script, the execution of the UPDATE will not be able to find the Table.
The solution is to separate these two commands into separate queries.