Fix for #133
InvalidOperationException: ExecuteReader requires the command to have a
transaction when the connection assigned to the command is in a pending
local transaction. The Transaction property of the command has not been
initialized.

We had the pending page transaction when we started a command for the
slug configuration. Fix was to send over the transaction if we have one.