Fixed javascript syntax issue in editor file.
Added developer version of piranha.js that will be injected into the manager in a developer environment and enable use of Vue tools.

Corrected TypeError being thrown by Vue.js developer version because apparently you aren't supposed to implicitly call global variables. Added piranha as a property on the window and Vue.prototype objects.

Added <base> element in the manager's _Layout file in case anyone wants to make use of document.baseURI. Currently points to an absolute path for the Url.

Removed a summernote script that was left as an unneeded dependency for CodeMirror.