Cache level and minutes headers added to response
This will return the cache level Piranha is using as well as the number of minutes until cache expires. This is helpful for those running on multiple instances especially if teams are working on Piranha together. There is not always a consistent response and the no-cache headers are deceiving, so adding these specific headers is informative and leads people to the solution to a somewhat common problem.